Refactored HelmVersion and GitVersion into separate classes.
